\name{add3}
\alias{add3}
\title{Give me an awsome title, please...}
\description{
  Give me an awsome description, please..
  }
\usage{
    add3(x,y,z)
}
\arguments{
  \item{x}{numeric, the 1st input}    
  \item{y}{numeric, the 2nd input}    
  \item{z}{numeric, the 3rd input}    
}
\value{
  The returned value is the sum of the three inputs
  
}
\examples{

library(osu)
add3(1,2,3)

}

\keyword{misc}
