\name{lorenz}
\alias{lorenz}
\title{Give me an awsome title, please... }
\description{
  Give me some awsome descriptions, please...
  }
\usage{
    lorenz(n=5000)
}
\arguments{
  \item{n}{integer, defaulted to 5000; the number of iterations}    
 
}
\value{
   Give me an awsome description, please..
  
}
\examples{

library(osu)
lorenz(n=5000)

}

\keyword{misc}
