\name{max}
\alias{max}
\title{Find the maximum of a vector}
\description{
  Give me an awsome description, please..
  }
\usage{
    mymax(x )
}
\arguments{
  \item{x}{a numeric vector}    
 
}
\value{
   Give me an awsome description, please..
  
}
\examples{

library(osu)
mymax( 1:10)

}

\keyword{misc}
