\name{myplot}
\alias{myplot}
\title{Give me an awsome title, please...}
\description{
  Give me an awsome description, please..
  }
\usage{
    myplot(vecNum)
}
\arguments{
  \item{vecNum}{ the input must be a vector of real nubmers.}    
}
\value{
  The returned value is a ggplot2 func.
  
}
\examples{

library(osu)
myplot(1:100)

}

\keyword{misc}
