\name{quad}
\alias{quad}
\title{A qudaratic equation solver}
\description{
  Find the roots of a quadratic equation from three coefficents
  }
\usage{
    quad(a,b,c)
}
\arguments{
  \item{x}{a numeric vector}    
 
}
\value{
   Give me an awsome description, please..
  
}
\examples{

library(osu)
quad(23,344,33)

}

\keyword{misc}
